@@ -1,12 +1,11 @@
 package com.example.eunice.cryptconvert.data;
 
-import android.arch.lifecycle.LiveData;
-import android.arch.lifecycle.MutableLiveData;
-import android.support.annotation.NonNull;
+import androidx.lifecycle.LiveData;
+import androidx.lifecycle.MutableLiveData;
+import androidx.annotation.NonNull;
 import android.util.Log;
 
 import com.example.eunice.cryptconvert.data.db.CryptoCurrency;
-import com.example.eunice.cryptconvert.data.network.ApiClient;
 import com.example.eunice.cryptconvert.data.network.ApiInterface;
 
 import retrofit2.Call;
