@@ -1,9 +1,9 @@
 package com.example.eunice.cryptconvert.ui;
 
 import android.app.Application;
-import android.arch.lifecycle.AndroidViewModel;
-import android.arch.lifecycle.LiveData;
-import android.support.annotation.NonNull;
+import androidx.lifecycle.AndroidViewModel;
+import androidx.lifecycle.LiveData;
+import androidx.annotation.NonNull;
 
 import com.example.eunice.cryptconvert.data.db.Country;
 import com.example.eunice.cryptconvert.data.db.CryptoCurrency;
