@@ -1,8 +1,7 @@
 package com.example.eunice.cryptconvert.data;
 
-import android.arch.lifecycle.LiveData;
+import androidx.lifecycle.LiveData;
 
-import com.example.eunice.cryptconvert.data.db.Country;
 import com.example.eunice.cryptconvert.data.db.CryptoCurrency;
 
 
