@@ -1,6 +1,6 @@
 package com.example.eunice.cryptconvert.data.repository;
 
-import android.arch.lifecycle.LiveData;
+import androidx.lifecycle.LiveData;
 
 import com.example.eunice.cryptconvert.data.db.Country;
 import com.example.eunice.cryptconvert.data.db.CryptoCurrency;
